Convert server to expose Tools following standard MCP formats and protocols
